fix: collapsible test
rename Sidebar.test.js to Collapsible.test.js
this also fixes the failing test in Sidebar
